Added 10MHz config to Clk1 and Clk2
In 1 new function, Clk1 is configured for external 10Mhz clock source and then Clk2 is configure to output 10Mhz, useful for chaining the 10Mhz on to a transverter or test equipment.  This appears to not be permanent so much be run before or sometime during SDR applications.  I do that with a batch file for each SDR app startup.